chore(email): move sendable email message
